requirements.txt
numpy==1.15.3
tensorflow==2.0.0-alpha0
keras==2.2.4
matplotlib==3.0.3
opencv-python==4.1.0.25
split_folders
scipy==1.1.0 
pandas==0.23.4
matplotlib==3.0.3
Cython
ipykernel
jupyter
path.py
Pillow
pygments
six
sphinx
wheel
zmq
requests
sklearn
